MODINV-291: Add 'Lost and paid' status for item.
