+++ descriptors/ModuleDescriptor-template.json
@@ -4,7 +4,7 @@
   "provides": [
     {
       "id": "inventory",
-      "version": "10.4",
+      "version": "10.5",
       "handlers": [
         {
           "methods": ["GET"],
@@ -289,7 +289,7 @@
   "requires": [
     {
       "id": "item-storage",
-      "version": "8.3"
+      "version": "8.4"
     },
     {
       "id": "instance-storage",
