+++ ramls/inventory.raml
@@ -1,6 +1,6 @@
 #%RAML 1.0
 title: Inventory API
-version: v10.4
+version: v10.5
 protocols: [ HTTP, HTTPS ]
 baseUri: http://localhost
