+++ ramls/item.json
@@ -258,7 +258,8 @@
             "Declared lost",
             "Order closed",
             "Claimed returned",
-            "Withdrawn"
+            "Withdrawn",
+            "Lost and paid"
           ]
         },
         "date": {
