+++ src/main/java/org/folio/inventory/domain/items/ItemStatusName.java
@@ -18,7 +18,8 @@ public enum ItemStatusName {
   DECLARED_LOST("Declared lost"),
   ORDER_CLOSED("Order closed"),
   CLAIMED_RETURNED("Claimed returned"),
-  WITHDRAWN("Withdrawn");
+  WITHDRAWN("Withdrawn"),
+  LOST_AND_PAID("Lost and paid");
 
   private static final Map<String, ItemStatusName> VALUE_TO_INSTANCE_MAP = initValueToInstanceMap();
   private final String value;
