+++ src/test/java/api/items/ItemAllowedStatusesSchemaTest.java
@@ -1,35 +1,24 @@
 package api.items;
 
-import static api.support.InstanceSamples.smallAngryPlanet;
 import static java.nio.file.Files.readAllBytes;
 import static java.nio.file.Paths.get;
-import static org.hamcrest.Matchers.is;
 import static org.junit.Assert.assertEquals;
-import static org.junit.Assert.assertThat;
 import static org.junit.Assert.assertTrue;
 
 import java.io.IOException;
-import java.net.MalformedURLException;
 import java.nio.charset.StandardCharsets;
 import java.util.Arrays;
 import java.util.Set;
-import java.util.UUID;
-import java.util.concurrent.ExecutionException;
-import java.util.concurrent.TimeoutException;
 import java.util.stream.Collectors;
 
 import org.folio.inventory.domain.items.ItemStatusName;
-import org.folio.inventory.support.http.client.IndividualResource;
 import org.junit.Test;
 import org.junit.runner.RunWith;
 
 import api.support.ApiTests;
-import api.support.builders.HoldingRequestBuilder;
-import api.support.builders.ItemRequestBuilder;
 import io.vertx.core.json.JsonArray;
 import io.vertx.core.json.JsonObject;
 import junitparams.JUnitParamsRunner;
-import junitparams.Parameters;
 
 /**
  * This test verifies that ramls/item.json and {@link ItemStatusName} is consistent.
@@ -51,32 +40,6 @@ public class ItemAllowedStatusesSchemaTest extends ApiTests {
 
     assertEquals("Schema enum does not match ItemStatusName values",
       enumAllowedItemStatuses, schemaAllowedItemStatuses);
-  }
-
-  @Test
-  @Parameters(method = "getItemStatusNameEnumAllowedItemStatuses")
-  public void canCreateItemsWithAllStatuses(String itemStatus) throws Exception {
-    final UUID holdingsId = createInstanceAndHolding();
-
-    final IndividualResource createResponse = itemsClient.create(
-      new ItemRequestBuilder()
-        .forHolding(holdingsId)
-        .canCirculate()
-        .withStatus(itemStatus));
-
-    assertThat(createResponse.getJson().getJsonObject("status").getString("name"),
-      is(itemStatus));
-  }
-
-  private UUID createInstanceAndHolding() throws MalformedURLException,
-    InterruptedException, ExecutionException, TimeoutException {
-
-    final IndividualResource instance = instancesClient
-      .create(smallAngryPlanet(UUID.randomUUID()));
-
-    return holdingsStorageClient
-      .create(new HoldingRequestBuilder().forInstance(instance.getId()))
-      .getId();
   }
 
   private Set<String> getSchemaAllowedItemStatuses() throws IOException {
